api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
httpx==0.24.1
redis==4.6.0
pydantic==2.3.0
python-dotenv==1.0.0
asyncio==3.4.3
aioredis==2.0.1
fastapi-cache2==0.2.1
pydantic-settings==2.0.3